refactor: simplify AdminCinemaWidgets by removing hardcoded cinema data and replacing the table with TableCard component
